Make Exporter an abstract base class
Also create implementations of Exporter:
 - FullExporter (corresponds to old Exporter)
 - EssentialExporter (corresponds to old GraphExporter)

Ref https://github.com/quintel/internal/issues/5#issuecomment-265103928
Ref https://github.com/quintel/internal/issues/5#issuecomment-265120551
